Safari-609.1.11.1/LayoutTests/media/media-source/media-source-delaying-load-event.html

<!DOCTYPE html>
<html>
<head>
    <title>media-source-delaying-load-event</title>
    <script src="../video-test.js"></script>
    <script>
    var source;
    </script>
</head>
<body>
    <div>A media element backed by a Media Source should not delay the window's load event</div>
    <video></video>
    <script>
        findMediaElement();

        source = new MediaSource();
        testExpected('source.readyState', 'closed');

        run('video.src = URL.createObjectURL(source)');
        waitForEventOn(window, 'load', endTest);
    </script>
</body>
</html>
