Safari-609.1.11.1/LayoutTests/media/media-source/media-source-delaying-load-event-expected.txt

A media element backed by a Media Source should not delay the window's load event

EXPECTED (source.readyState == 'closed') OK
RUN(video.src = URL.createObjectURL(source))
EVENT(load)
END OF TEST
